Make English the default language for the thesis
Currently Dutch is the default language for the thesis.
Various packages will use Dutch to generate strings such as "last accessed on"
for references in Dutch.
That is the opposite of what we want.

See https://www.overleaf.com/learn/latex/International_language_support
to see why the order matters.